Convert type early

+++ ledger/block/src/transaction/deployment/mod.rs
@@ -126,7 +126,7 @@ impl<N: Network> Deployment<N> {
 
     /// Returns the total number of constraints.
     pub fn num_constraints(&self) -> u64 {
-        self.verifying_keys.iter().map(|(_, (vk, _))| vk.circuit_info.num_constraints).sum::<usize>() as u64
+        self.verifying_keys.iter().map(|(_, (vk, _))| vk.circuit_info.num_constraints as u64).sum::<u64>()
     }
 
     /// Returns the deployment ID.
